added_skip_camera_init
it is in cli form so you can change the bat file cli to --skip-remap
Sometimes the axis, either x or y, flips on camera initialization, and that may contradict the old prealigned xmp files.  This is my guess what causes the model to rotate on the bank a little bit. Lets fix that on wednesday.